Refactor project structure
- Divide into 3 modules: server (main), handler, and protocol
- 2 types of handler and protocol: normal client and VIP
